MissingCover/requirements.txt
Flask==3.1.1
requests==2.32.3
Pillow==10.4.0
openai==1.58.1
tqdm==4.66.5
scikit-learn==1.6.1
opencv-python==4.11.0.86
ultralytics==8.3.139
vllm==0.7.3 
torch==2.5.1+cu118
transformers==4.49.0